+++ crates/vm/helpers.rs
@@ -1,4 +1,5 @@
 use ethrex_common::types::{ChainConfig, Fork};
+use ethrex_trie::{Node, NodeRLP, PathRLP, Trie};
 pub use revm::primitives::SpecId;
 
 /// Returns the spec id according to the block timestamp and the stored chain config
@@ -41,4 +42,46 @@ pub fn create_contract_address(from: Address, nonce: u64) -> Address {
             .0
             .as_ref(),
     )
+}
+
+/// Get all potential child nodes of a node whose value was deleted.
+///
+/// After deleting a value from a (partial) trie it's possible that the node containing the value gets
+/// replaced by its child, whose prefix is possibly modified by appending some nibbles to it.
+/// If we don't have this child node (because we're modifying a partial trie), then we can't
+/// perform the deletion. If we have the final proof of exclusion of the deleted value, we can
+/// calculate all posible child nodes.
+pub fn get_potential_child_nodes(proof: &[NodeRLP], key: &PathRLP) -> Option<Vec<Node>> {
+    // TODO: Perhaps it's possible to calculate the child nodes instead of storing all possible ones.
+    let trie = Trie::from_nodes(
+        proof.first(),
+        &proof.iter().skip(1).cloned().collect::<Vec<_>>(),
+    )
+    .unwrap();
+
+    // return some only if this is a proof of exclusion
+    if trie.get(key).unwrap().is_none() {
+        let final_node = Node::decode_raw(proof.last().unwrap()).unwrap();
+        match final_node {
+            Node::Extension(mut node) => {
+                let mut variants = Vec::with_capacity(node.prefix.len());
+                while {
+                    variants.push(Node::from(node.clone()));
+                    node.prefix.next().is_some()
+                } {}
+                Some(variants)
+            }
+            Node::Leaf(mut node) => {
+                let mut variants = Vec::with_capacity(node.partial.len());
+                while {
+                    variants.push(Node::from(node.clone()));
+                    node.partial.next().is_some()
+                } {}
+                Some(variants)
+            }
+            _ => None,
+        }
+    } else {
+        None
+    }
 }
